% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_0_hypergrid.R
\name{build_hyperparameter_dataframe}
\alias{build_hyperparameter_dataframe}
\title{Creation of a hyperparameter dataset to use in ranger function to optimize a random forest model}
\usage{
build_hyperparameter_dataframe(
  mtry_frac = NULL,
  min.node.size = NULL,
  sample.fraction = NULL,
  ntrees = NULL,
  weight_list = weight_list,
  PEMs = NULL,
  phylo
)
}
\arguments{
\item{mtry_frac}{numeric vector of values}

\item{min.node.size}{numeric vector of values}

\item{sample.fraction}{numeric vector of values}

\item{ntrees}{numeric vector of values}

\item{phylo}{TRUE or FALSE}

\item{wgt}{numeric vector of values}

\item{PEMS}{numeric vector of values}
}
\description{
Creation of a hyperparameter dataset to use in ranger function to optimize a random forest model
}
