% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Data: Species x Traits and Phylogeny}
\format{
A \code{data.frame} with 200 rows (species) and the following two
columns:
\describe{
\item{Species_ID}{code to identify species, each species must be unique}
\item{Numeric_Trait_X}{value of the numeric trait number X for a unique species}
\item{Categorical_Trait_X}{category of the categorical trait number X for a unique species}
\item{EigX}{values of the eigenvector number X for a unique species}
}
}
\usage{
dataset
}
\description{
This dataset is a generic dataset corresponding to species for which we
various traits, continuous and categorical and phylogenetic eigenvectors.
The last column is a response variable, here the number of predation events
per year per species.
}
\examples{
'Need to be done'
}
\keyword{datasets}
