% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_RF_optimization.R
\name{optimized_RF_function}
\alias{optimized_RF_function}
\title{opitmized_RF_function}
\usage{
optimized_RF_function(
  data,
  species,
  response_data,
  trait_data,
  phylo_data = NULL,
  classification,
  weight = NULL,
  mtry_frac = NULL,
  min.node.size = NULL,
  sample.fraction = NULL,
  ntrees = NULL,
  PEMs = NULL,
  cores = 1
)
}
\arguments{
\item{data}{A species_traits dataframe containing species traits and eigen values if the user use phylogenetic data}

\item{species}{The name of the column corresponding to the species names in the @param data}

\item{response_data}{The name of the column corresponding to the response in the @param data}

\item{trait_data}{A dataframe containing only species trait data}

\item{phylo_data}{By default, phylo_data = NULL, which return an expand grid without PEMs}

\item{classification}{Can be eiter TRUE or FALSE. If TRUE, you want to perform a classification analyses, if FALSE, you want to perform a regression analyses}

\item{weight}{A list containing vector as the same legnth of row in the data}

\item{mtry_frac}{By default, NULL, return a list of default values of mtry. Else, you define your own range of mtry value you want to optimize in a vector}

\item{min.node.size}{By default, NULL, return a list of default values of minimum node size. Else, you define your own range of minimum node size value you want to optimize in a vector}

\item{sample.fraction}{By default, NULL, return a list of default values of fraction of observations to sample. Else, you define your own range of observations to sample value you want to optimize in a vector}

\item{ntrees}{By default, NULL, return a list of default values of number of trees. Else, you define your own range of number of trees you want to optimize in a vector}

\item{cores}{Define the number of cores you want to use. By default, the number of cores is set to 1. For Windows user, keep number of cores to 1, or switch to linux/Mac OS.}
}
\value{
Return a dataframe
}
\description{
opitmized_RF_function
}
