% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_RF_quality.R
\name{random_forest_quality}
\alias{random_forest_quality}
\title{Computing metric of random forest quality based on the training data used for optimization}
\usage{
random_forest_quality(Yobs, Ypred, method, threshold = NULL)
}
\arguments{
\item{Yobs}{numeric vector}

\item{Ypred}{numeric vector}

\item{method}{'regression' or 'classification'}

\item{threshold}{numeric value}
}
\description{
Computing metric of random forest quality based on the training data used for optimization
}
