% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_1_weigthing.R
\name{weight_scenarii_list}
\alias{weight_scenarii_list}
\title{Creation of scenarii to weight the observation for modelisation with random forest}
\usage{
weight_scenarii_list(Y_obs, weight_for_non_prey = NULL)
}
\arguments{
\item{Y_obs}{numeric vector}

\item{weight_for_non_prey}{numeric value}
}
\description{
Creation of scenarii to weight the observation for modelisation with random forest
}
