% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_0_hypergrid.R
\name{build_hyperparameter_dataframe}
\alias{build_hyperparameter_dataframe}
\title{Building an hyperparameter dataset to optimized random forest using ranger from ranger}
\usage{
build_hyperparameter_dataframe(
  mtry_frac = NULL,
  min.node.size = NULL,
  sample.fraction = NULL,
  ntrees = NULL,
  weight_list = weight_list,
  PEMs = NULL,
  phylo
)
}
\arguments{
\item{mtry_frac}{a numeric vector}

\item{min.node.size}{a numeric vector}

\item{sample.fraction}{a numeric vector}

\item{ntrees}{a numeric vector}

\item{weight_list}{a list of numeric vector}

\item{PEMs}{a numeric vector}

\item{phylo}{a Boolean TRUE or FALSE}
}
\value{
a dataframe which contain all Random Forest option combination
}
\description{
Building an hyperparameter dataset to optimized random forest using ranger from ranger
}
\examples{

\dontrun{

# First load the data

library(predipreda)
data(dataset)

# The build_hyperparameter_dataframe function depends on the weight_scenarii_list function to generate a weight list

weight_list <- weight_scenarii_list(dataset$Response)

# Then, run the build_hyperparameter_dataframe function which return a dataframe. Each line of the dataframe is a combination of option for the Random Forest
hyper_grid <- build_hyperparameter_dataframe(mtry_frac=NULL,
                                             min.node.size=NULL,
                                             sample.fraction=NULL,
                                             ntrees=NULL,
                                             weight_list=weight_list,
                                             PEMs=NULL,
                                             phylo = TRUE)

}
}
