% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_1_weigthing.R
\name{weight_scenarii_list}
\alias{weight_scenarii_list}
\title{Creation of scenarii to weight the observation for modelisation with random forest}
\usage{
weight_scenarii_list(Y_obs, weight_for_non_prey = NULL)
}
\arguments{
\item{Y_obs}{numeric vector}

\item{weight_for_non_prey}{numeric value}
}
\value{
a list of scenarii. Each scenario is a vector of weights
}
\description{
Creation of scenarii to weight the observation for modelisation with random forest
}
\examples{
\dontrun{

# First load the data

library(predipreda)
data(dataset)

# The weight_scenarii_list function generate a weight list. Each object of the list is a scenario, which is a vector of length equal to the lenght of the response. Hence, each object of the vector is a weight.

weight_list <- weight_scenarii_list(dataset$Response)
}
}
